requirements.txt
python-dotenv
pydantic
click
pika
rich